+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-### Unreleased
+### 3.10.0
 - Add `-Images` parameter to `Add-ThreadMessage`.  
   This makes that the image input can be used in Assistants.
 - Fix issues about relative file path input.